fast/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
beautifulsoup4==4.12.3
click==8.1.7
diskcache==5.6.3
fastapi==0.115.2
greenlet==3.0.3
h11==0.14.0
htmlmin==0.1.12
idna==3.10
pillow==11.0.0
playwright==1.47.0
pydantic==2.9.2
pydantic_core==2.23.4
pyee==12.0.0
python-dotenv==1.0.1
python-multipart==0.0.12
sniffio==1.3.1
soupsieve==2.6
starlette==0.40.0
typing_extensions==4.12.2
uvicorn==0.32.0
